requirements-tg.txt
asyncpg==0.31.0
aioboto3==15.5.0
openai==2.14.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
aiogram==3.20.0
babel==2.17.0
rembg[cpu]==2.0.69
pycryptodome==3.23.0